requirements.txt
# Runtime dependencies
neo4j==5.28.1
colorama==0.4.6
typeguard==4.4.3
python-dotenv==1.1.0

# Testing dependencies  
pytest==8.4.0